requirements.txt
numpy==1.23.5
Flask==3.1.0
pandas==1.5.3
python-dotenv==1.1.0
pyodbc==5.2.0